QCamera2: HAL1: Bug fix in dequeuing the deferred preview buffers
Issue:
Preview Buffer Memory leak due to mapping more than once if the same dequeued
Buffer Index comes before finishing the dequeueing of all the deferred buffers.
For eg:
Assume Total preview buffer count is 20 and initially we map only 5 preview
and remaining all 15 are dequeued in the First preview CB.
Initially mMappableBuffers is 5 and by the time it dequeue's 8 buffers assume same
buffer index of 6 queued to daemon -> came back to HAL -> sent to display through
sync CB -> display returned buffer and then again HAL tries to map buffer since
numMapped is not dynamically updated.
It's because For Loop count is based on getMappable() but the corresponding
mMappableBuffers is incremented in dequeueBuffer() which is not taken care.

Fix:
Update the numMapped dynamically so that same buffer won't be mapped again.

CRs-Fixed: 1064419

Change-Id: Ie0ccc1e2c7df39eb3e19334fddefbd83bf88f82f