QCamera2: Add lock protection for stream list access.
Issue:
In liveshot usecase, linked streams are added in a new thread
leading to modification of mStreams list in channel. This can
lead to race conditions and/or crash if the same list is accessed
from a different thread.

In the issue case, crash was seen in zoom event, trying to set
crop info in all the streams, being called from state machine
thread context and the livesnapshot thread trying to add linked
stream at the same time.

Fix:
Use lock to protect modification/access of mStreams.

Change-Id: I88ce793f1756a546fac84cbb16130d4a70beb9a9
CRs-Fixed: 1058021
Change-Id: I83e02e640424ca8b301f03f0541d18e3b2ad50e3
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>